+++ core/src/com/mikroysoft/game/Coordinate.java
@@ -1,10 +1,10 @@
 package com.mikroysoft.game;
 
 public class Coordinate {
-	public int x;
-	public int y;
-	public Coordinate(int x, int y) {
-		this.x = x;
-		this.y = y;
-	}
+    public float x;
+    public float y;
+    public Coordinate(float x, float y) {
+        this.x = x;
+        this.y = y;
+    }
 }
